requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.116.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
joblib==1.5.1
keras==3.11.1
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.1.0
numpy==1.24.4
opt_einsum==3.4.0
optree==0.17.0
openai
packaging==25.0
protobuf==4.25.8
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.1
tensorflow-io-gcs-filesystem
termcolor==3.1.0
tf_keras==2.17.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
wrapt==1.17.2
